Revert "Move UnitializedClient to server"
This reverts commit 0cbe063c77500f607ad653cced8a3782568bc816.

Attempt to fix https://github.com/facebook/flipper/runs/3351826251?check_suite_focus=true
